Update new twake welcome screen
(cherry picked from commit bdda9720904720808b804edf10ded9df60c7534d)
