+++ lib/config/go_routes/go_router.dart
@@ -12,7 +12,6 @@ import 'package:fluffychat/pages/chat_draft/draft_chat_adaptive_scaffold.dart';
 import 'package:fluffychat/pages/chat_encryption_settings/chat_encryption_settings.dart';
 import 'package:fluffychat/pages/error_page/error_page.dart';
 import 'package:fluffychat/pages/homeserver_picker/homeserver_picker.dart';
-import 'package:fluffychat/pages/twake_id/twake_id.dart';
 import 'package:fluffychat/pages/new_group/new_group_chat_info.dart';
 import 'package:fluffychat/pages/settings_dashboard/settings_app_language/settings_app_language.dart';
 import 'package:fluffychat/pages/settings_dashboard/settings_profile/settings_profile.dart';
@@ -60,7 +59,7 @@ abstract class AppRoutes {
     BuildContext context,
     GoRouterState state,
   ) =>
-      Matrix.of(context).client.isLogged() ? null : '/home/twakeid';
+      Matrix.of(context).client.isLogged() ? null : '/home/twakeWelcome';
 
   AppRoutes();
 
@@ -91,10 +90,10 @@ abstract class AppRoutes {
           redirect: loggedInRedirect,
         ),
         GoRoute(
-          path: 'twakeid',
+          path: 'twakeWelcome',
           pageBuilder: (context, state) => defaultPageBuilder(
             context,
-            const TwakeId(),
+            const TwakeWelcome(),
           ),
           redirect: loggedInRedirect,
         ),
@@ -365,10 +364,10 @@ abstract class AppRoutes {
                   redirect: loggedOutRedirect,
                 ),
                 GoRoute(
-                  path: 'twakeid',
+                  path: 'twakeWelcome',
                   pageBuilder: (context, state) => defaultPageBuilder(
                     context,
-                    const TwakeId(),
+                    const TwakeWelcome(),
                   ),
                   redirect: loggedInRedirect,
                 ),
