+++ lib/pages/twake_welcome/twake_welcome_view_style.dart
@@ -1,17 +1,4 @@
-import 'package:flutter/material.dart';
-
 class TwakeWelcomeViewStyle {
-  static const double logoWidth = 257;
-  static const double logoHeight = 179;
-
-  static const double buttonHeight = 56;
-
-  static const double buttonRadius = 100;
-
-  static const EdgeInsets descriptionPadding =
-      EdgeInsets.symmetric(horizontal: 8);
-
-  static const EdgeInsets buttonPadding = EdgeInsets.only(
-    bottom: 44,
-  );
+  static const double logoWidth = 225;
+  static const double logoHeight = 147;
 }
